requirements.txt
absl-py
alembic
appnope
argon2-cffi
astor
async-generator
attrs
backports.functools-lru-cache
bleach
Brotli
brotlipy
certifi
cffi
chardet
chart-studio
click
colorlover
cryptography
decorator
defusedxml
entrypoints
Flask
Flask-Compress
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
future
gast
google-pasta
greenlet
grpcio
gunicorn==20.1.0
h5py
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
Jinja2
joblib
jsonschema
Keras-Applications
Keras-Preprocessing
Mako
Markdown
MarkupSafe
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
packaging
pandas
pandocfilters
patsy
pexpect
pickleshare
plotly
plotly-express
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pycparser
Pygments
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil
python-editor
pytz
pyzmq
requests
retrying
scikit-learn
scipy
Send2Trash
simplegeneric
six
SQLAlchemy
statsmodels
tensorboard
tensorflow-cpu
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
tornado
traitlets
typing-extensions
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
wrapt
WTForms
zipp
